requirements.txt
flask
pymysql
python-dotenv
openai 
fastapi 
uvicorn  
requests
gradio
